--- examples/cdk/aos-ddb-search-app/lib/appsync/codegen/graphql/queries.ts
@@ -1,34 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// this is an auto generated file. This will be overwritten
-
-export const getTodo = /* GraphQL */ `
-  query GetTodo($id: ID!) {
-    getTodo(id: $id) {
-      id
-      title
-      description
-      owner
-      __typename
-    }
-  }
-`;
-export const listTodos = /* GraphQL */ `
-  query ListTodos(
-    $filter: TableTodoFilterInput
-    $limit: Int
-    $nextToken: String
-  ) {
-    listTodos(filter: $filter, limit: $limit, nextToken: $nextToken) {
-      items {
-        id
-        title
-        description
-        owner
-        __typename
-      }
-      nextToken
-      __typename
-    }
-  }
-`;
