+++ examples/cdk/aoss-search-app/.eslintrc.json
@@ -1,0 +1,29 @@
+{
+	"env": {
+		"es2021": true,
+		"node": true
+	},
+	"extends": ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+	"parser": "@typescript-eslint/parser",
+	"parserOptions": {
+		"ecmaVersion": "latest",
+		"sourceType": "module"
+	},
+	"plugins": ["@typescript-eslint"],
+	"rules": {},
+	"overrides": [
+		{
+			"files": ["lib/appsync/resolvers/*.ts"],
+			"extends": [
+				"eslint:recommended",
+				"plugin:@typescript-eslint/recommended",
+				"plugin:@aws-appsync/recommended"
+			],
+			"parserOptions": {
+				"ecmaVersion": "latest",
+				"sourceType": "module",
+				"project": "./tsconfig.json"
+			}
+		}
+	]
+}
