+++ examples/cdk/aoss-search-app/.gitignore
@@ -6,3 +6,7 @@ node_modules
 # CDK asset staging directory
 .cdk.staging
 cdk.out
+
+**/codegen/graphql
+**/codegen/API.ts
+output.json
