+++ examples/cdk/aoss-search-app/.npmignore
