+++ examples/cdk/aoss-search-app/README.md
@@ -1,0 +1,73 @@
+# Todo Search API - CDK Example
+
+This CDK app provides an implementation for a simple Todo API  with AppSync JavaScript resolvers.
+
+The API allows you to interact with a Todo type that is backed by a DynamoDB table.
+
+```graphql
+type Todo {
+  id: ID!
+  title: String
+  description: String
+  owner: String
+}
+```
+
+The app uses the [AppSync Helper construct](../constructs/appsync-helper/) for a quick setup.
+
+```sh
+lib
+├── appsync
+│   ├── codegen
+│   │   └── index.ts
+│   ├── resolvers
+│   │   ├── Mutation.createTodo.[todos].ts
+│   │   ├── Mutation.deleteTodo.[todos].ts
+│   │   ├── Mutation.updateTodo.[todos].ts
+│   │   ├── Query.getTodo.[todos].ts
+│   │   ├── Query.listTodos.[todos].ts
+│   │   └── utils.ts
+│   └── schema.graphql
+├── appsync-helper.ts
+└── dynamodb-todo-app-stack.ts
+```
+
+The [appsync](./lib/appsync-helper.ts) folder contains schema and the resolver code. This project uses unit resolvers, but you can expand it to use pipeline resolvers if needed.
+
+## Init
+
+Install all the dependencies. From the [project folder](./)
+
+```sh
+npm install
+```
+
+## Codegen
+
+This TypeScript projects uses types generated by [Amplify Codegen](https://docs.amplify.aws/cli-legacy/graphql-transformer/codegen/). To populate your codegen and generate it after you update your schema, change directory to [appsync](./lib/appsync/) and run:
+
+```sh
+npx amplify codegen
+```
+
+## Deploy the stack
+
+To deploy the stack, from the top folder:
+
+```sh
+npm run cdk deploy -- -O output.json
+```
+
+Once deployed, you can find your API **TodoAPI** in the AWS console.
+
+## Connect to an app
+
+You can use the info in [./output.json](./output.json) to configure your Amplify client in your app. See [Building a client application](https://docs.aws.amazon.com/appsync/latest/devguide/building-a-client-app.html).
+
+## Destroy the stack
+
+To destroy the stack and all the created resources:
+
+```sh
+npm run cdk destroy
+```
