+++ examples/cdk/aoss-search-app/jest.config.js
