+++ examples/cdk/aoss-search-app/lib/appsync-helper.ts
