+++ examples/cdk/aoss-search-app/lib/appsync/.eslintrc.json
