+++ examples/cdk/aoss-search-app/lib/appsync/.graphqlconfig.yml
