+++ examples/cdk/aoss-search-app/lib/appsync/codegen/index.ts
