+++ examples/cdk/aoss-search-app/lib/appsync/resolvers/Query.search.[aoss].ts
@@ -1,0 +1,39 @@
+import { Context } from '@aws-appsync/utils';
+import { search } from './utils';
+import { Result, SearchQueryVariables, SearchableTodoConnection } from '@codegen';
+
+export function request(ctx: Context<SearchQueryVariables>) {
+	const { filter, limit, nextToken } = ctx.args;
+	const match_all = JSON.stringify({ match_all: {} });
+	const bq = !filter
+		? match_all
+		: (util.transform.toElasticsearchQueryDSL(ctx.args.filter as any) as unknown as string);
+	const body: Record<string, unknown> = { query: JSON.parse(bq) };
+	if (nextToken) {
+		body.search_after = JSON.parse(util.base64Decode(nextToken));
+	}
+	return search({
+		index: 'todos',
+		body,
+		size: limit ?? 100,
+		sort: 'id.keyword:asc',
+	});
+}
+
+export function response(ctx: Context): Result<SearchableTodoConnection> {
+	const body = JSON.parse(ctx.result.body);
+	if (!`${ctx.result.statusCode}`.startsWith('2')) {
+		util.error(body.error?.reason, body.error?.type, body);
+	}
+	let nextToken = null;
+	const hits = body.hits.hits;
+	const hl = hits.length;
+	if (hl > 0) {
+		nextToken = util.base64Encode(JSON.stringify(hits[hl - 1].sort));
+	}
+	return {
+		items: hits.map((hit: any) => hit._source),
+		total: body.hits.total.value,
+		nextToken,
+	};
+}
