+++ examples/cdk/aoss-search-app/package.json
@@ -1,8 +1,8 @@
 {
-	"name": "aos-ddb-search-app",
+	"name": "aoss-search-app",
 	"version": "0.1.0",
 	"bin": {
-		"aos-ddb-search-app": "bin/aos-ddb-search-app.js"
+		"aoss-search-app": "bin/aoss-search-app.js"
 	},
 	"scripts": {
 		"build": "tsc",
@@ -11,14 +11,15 @@
 		"cdk": "cdk"
 	},
 	"devDependencies": {
-		"@aws-appsync/eslint-plugin": "^1.2.1",
 		"@aws-appsync/utils": "^1.2.5",
 		"@graphql-tools/schema": "^10.0.0",
 		"@types/jest": "^29.5.1",
 		"@types/node": "20.1.7",
 		"@typescript-eslint/eslint-plugin": "^6.3.0",
+		"@typescript-eslint/parser": "^6.3.0",
 		"aws-cdk": "2.87.0",
 		"esbuild": "^0.18.17",
+		"eslint": "^8.46.0",
 		"graphql": "^16.7.1",
 		"jest": "^29.5.0",
 		"ts-jest": "^29.1.0",
@@ -26,7 +27,6 @@
 		"typescript": "~5.1.3"
 	},
 	"dependencies": {
-		"@aws-appsync/utils": "^1.2.5",
 		"aws-cdk-lib": "2.87.0",
 		"constructs": "^10.0.0",
 		"source-map-support": "^0.5.21"
