+++ examples/cdk/aoss-search-app/test/aos-ddb-search-app.test.ts
@@ -1,0 +1,16 @@
+// import * as cdk from 'aws-cdk-lib';
+// import { Template } from 'aws-cdk-lib/assertions';
+// import * as AosDdbSearchApp from '../lib/aoss-search-app-stack';
+
+// example test. To run these tests, uncomment this file along with the
+// example resource in lib/aoss-search-app-stack.ts
+test('SQS Queue Created', () => {
+	//   const app = new cdk.App();
+	//     // WHEN
+	//   const stack = new AosDdbSearchApp.AosDdbSearchAppStack(app, 'MyTestStack');
+	//     // THEN
+	//   const template = Template.fromStack(stack);
+	//   template.hasResourceProperties('AWS::SQS::Queue', {
+	//     VisibilityTimeout: 300
+	//   });
+});
