+++ examples/cdk/aoss-search-app/tsconfig.json
