+++ examples/cdk/dynamodb-todo-app/package.json
@@ -27,7 +27,6 @@
 		"typescript": "~5.1.3"
 	},
 	"dependencies": {
-		"@aws-appsync/eslint-plugin": "file:../../../../yuth/aws-appsync-toolkit/packages/eslint-plugin-appsync",
 		"aws-cdk-lib": "2.87.0",
 		"constructs": "^10.0.0",
 		"source-map-support": "^0.5.21"
